SkillsOverview refactored und neuen Service erstellt.
